+++ crates/integrations/connector-integration/src/connectors/adyen/test.rs
@@ -83,8 +83,8 @@ mod tests {
                     connector_response: None,
                     recurring_mandate_payment_data: None,
                     l2_l3_data: None,
-
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
 },
                 connector_config: ConnectorSpecificConfig::Adyen {
                     api_key: Secret::new(api_key),
@@ -282,8 +282,8 @@ mod tests {
                     connector_response: None,
                     recurring_mandate_payment_data: None,
                     l2_l3_data: None,
-
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
 },
                 connector_config: ConnectorSpecificConfig::Adyen {
                     api_key: Secret::new(api_key),