+++ crates/integrations/connector-integration/src/connectors/calida/test.rs
@@ -106,6 +106,7 @@ mod tests {
                     recurring_mandate_payment_data: None,
                     l2_l3_data: None,
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Calida {
                     api_key: Secret::new(api_key),
@@ -266,6 +267,7 @@ mod tests {
                     recurring_mandate_payment_data: None,
                     l2_l3_data: None,
                     merchant_request_id: None,
+                    sender_payment_instrument_id: None,
                 },
                 connector_config: ConnectorSpecificConfig::Calida {
                     api_key: Secret::new(api_key),