+++ examples/bamboraapac/bamboraapac.py
@@ -16,9 +16,14 @@ SUPPORTED_FLOWS = ["authorize", "capture", "get", "proxy_authorize", "proxy_setu
 
 _default_config = sdk_config_pb2.ConnectorConfig(
     options=sdk_config_pb2.SdkOptions(environment=sdk_config_pb2.Environment.SANDBOX),
-    # connector_config=payment_pb2.ConnectorSpecificConfig(
-    #     bamboraapac=payment_pb2.BamboraapacConfig(api_key=...),
-    # ),
+    connector_config=payment_pb2.ConnectorSpecificConfig(
+        bamboraapac=payment_pb2.BamboraapacConfig(
+            username=payment_methods_pb2.SecretString(value="YOUR_USERNAME"),
+            password=payment_methods_pb2.SecretString(value="YOUR_PASSWORD"),
+            account_number=payment_methods_pb2.SecretString(value="YOUR_ACCOUNT_NUMBER"),
+            base_url="YOUR_BASE_URL",
+        ),
+    ),
 )